poll with a proper waker in `isyswasfa-guest`
Previously, we were using a dummy, no-op waker and keeping track of which
futures to poll out-of-band.  That works in most cases, but not when polling a
given future might cause other futures to become ready (e.g. channels,
`tokio::sync::Mutex`, etc.).  In the latter cases, we need the waker to tell us
which additional futures to poll.  That also means we need to iterate to a fixed
point when polling since each loop may generate more futures to poll.

Note that we're still using a dummy, no-op waker on the host side, but we can
get away with it there since we poll all leaf futures using a
`futures::stream::FuturesUnordered` which interposes its own wakers internally.
And for `wasi:io/poll/pollable`s, we create and poll a fresh future on every
loop, so there's nothing for a waker to wake once we're done.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>