eliminate need for `Mutex` in async host functions
Consider this WIT type and imagine we're implementing it in the host:
```
resource foo {
  bar: static func(x: foo) -> foo;
}
```

Previously, we generated a trait that looked something like this:
```
trait HostFoo: IsyswasfaView {
    async fn bar(state: Self::State, x: Resource<Foo>) -> Result<Resource<Foo>>;
    fn drop(&mut self, this: Resource<Foo>) -> Result<()>;
}
```
...where we used Self::State (which is `Send + Sync + 'static`, e.g. a `Arc<Mutex<T>>` or whatever) instead of `&mut self` so we could return back to the guest without blocking, but still give the implementation mutable access to a `ResourceTable`, etc.

However, I realized that we can avoid the reference counting, locking, and any extra allocation by recognizing that the implementation usually only needs access to e.g. the `ResourceTable` before and after it `await`s something.  So we can instead generate something like this:
```
trait HostFoo {
  fn bar(&mut self, x: Resource<Foo>) -> Result<impl Future<Output = impl FnOnce(&mut Self) -> Result<Resource<Foo>> + 'static> + Send + 'static>;
  fn drop(&mut self, this: Resource<Foo>) -> Result<()>;
}
```
That lets the implementation do what it likes with `&mut self`, then return a `Future` (likely just an `async move { ... }`) which can await as many times as it likes.  The output of that `Future` is a `FnOnce` which gets access to `&mut self` one more time and then finally returns the result.  For most implementations, this all that's needed for an optimum, reasonably ergonomic implementation.  There might still be cases where an implementation needs access to store data between `await`s, in which case it will need to fall back to some kind of `Arc<Mutex<T>>` that gets moved into the future, but I expect that will be quite rare in practice.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>